test(tui): co-locate app tests with their submodules (KAN-762)
